+++ src/main/java/com/cells/parts/AbstractCombinedInterfacePart.java
@@ -57,6 +57,7 @@ import com.cells.blocks.interfacebase.IInterfaceLogic;
 import com.cells.blocks.interfacebase.fluid.FluidInterfaceLogic;
 import com.cells.blocks.interfacebase.item.ItemInterfaceLogic;
 import com.cells.gui.CellsGuiHandler;
+import com.cells.helpers.InterfaceMemoryCardHelper;
 import com.cells.helpers.InterfaceApiHelper;
 import com.cells.integration.mekanismenergistics.CombinedInterfaceGasHelper;
 import com.cells.integration.mekanismenergistics.MekanismEnergisticsIntegration;
@@ -410,9 +411,13 @@ public abstract class AbstractCombinedInterfacePart extends PartBasicState
                 memoryCard.notifyUser(player, MemoryCardMessages.SETTINGS_SAVED);
             }
         } else {
-            final String storedName = memoryCard.getSettingsName(memCardIS);
-            final NBTTagCompound data = memoryCard.getData(memCardIS);
-            if (name.equals(storedName)) {
+            final NBTTagCompound data = InterfaceMemoryCardHelper.prepareUploadData(
+                memoryCard.getSettingsName(memCardIS),
+                memoryCard.getData(memCardIS),
+                InterfaceMemoryCardHelper.combined(this.isExport())
+            );
+
+            if (data != null) {
                 this.uploadSettings(SettingsFrom.MEMORY_CARD, data, player);
                 memoryCard.notifyUser(player, MemoryCardMessages.SETTINGS_LOADED);
             } else {